Be truthful with the compiler about alignment tricks
In the interest of performance, we copy and compare 6-byte ethernet
hardware addresses as a 4-byte `unsigned long` and a 2-byte `unsigned
short`. But the 68000 can only operate on those types when they are
aligned at on word boundaries, and throws an Address Error otherwise.

Instead of describing hardware addresses as just an array of bytes and
casting pointers willy nilly, describe them as a `union` of a byte array
and the `unsigned long`/`unsigned short` combo. This means that the
compiler will enforce alignment and all will be happy (and we should get
better performance on the 030 as a bonus).